Change floating link style and position
As the big floating link was appearing in bothering places when triple clicking, the new logic is to place it on top of the position where the double click event is captured.
In case the space to the top is less than the floating link height, it is placed below the same reference point.
Also the new style shrinks a bit the link and places an arrow below or above the rectangle depending on its position.